docs(slices): bring all core slices to Tier-1 documentation depth (#602)
* docs(slices): bring all core slices to Tier-1 documentation depth

Add advisory authoring metadata (gmeow:useWhen / avoidWhen / howToUse +
skos:example, plus useForConsumer/avoidForConsumer where projection-relevant)
to every class and object/datatype property across the core slices, clearing
the Tier-1 depth warning backlog (#433 cat-1, #325/#471): structural lint goes
from ~2072 Tier-1 warnings to 0. Also polishes manifest slice metadata and adds
per-term docs.md anchors (#325 Tier-2) to the guides that lacked them.

Fixes a regression where InstantMessageAddress was dropped during definition
polish (restored with full Tier-1 fields). Rewrites the graded-gate test to
exercise the warn-not-error invariant via the synthetic probe graph, since the
live backlog it relied on is now cleared.

Touches two extension slices (lexicon, norms) opportunistically.

* docs(slices): address PR review feedback

- contacts: add gmeow:consumerPublicSite to avoidForConsumer on ContactPoint
  and InstantMessageAddress, matching the sibling email/telephone/postal terms
  so sensitive contact channels are not projected to public sites (CodeRabbit).
- organization: declare gmeow:identifierValue and gmeow:identifierScheme
  owl:FunctionalProperty — one value + one scheme per reified gmeow:Identifier
  node, consistent with the already-functional quantityValue/monetaryValue/
  durationValue sibling pattern (Gemini).
- gts: fix stray punctuation "(spec §3.2, )" -> "(spec §3.2)" in the
  GTSCompaction definition (CodeRabbit).
- names: complete the PronounSet example to all five forms and add
  gmeow:usageAuthority to the PlaceNaming example so each snippet shows the
  axes its prose describes (CodeRabbit).

Regenerated; reason (OWL 2 DL + ELK consistency) green.

+++ src/gmeow_tools/ext/music/serializers/musicxml.py
@@ -26,7 +26,7 @@ def _pitch_elements(pitch: PitchValue) -> tuple[str, float, int]:
     Microtonal deviations are expressed as decimal ``alter`` values.
     """
     midi = pitch.to_midi_number()
-    rounded = int(round(midi))
+    rounded = round(midi)
     chroma = rounded % 12
     octave = (rounded // 12) - 1
     step, semitone_alter = [
